moved the upgrade code to pkg
